dozor_endpoint_security/requirements.txt
asn1crypto==1.5.1
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
decorator==5.1.1
defusedxml==0.7.1
dissect==3.10
dissect.cim==3.7
dissect.clfs==1.6
dissect.cstruct==3.10
dissect.esedb==3.9
dissect.etl==3.7
dissect.eventlog==3.6
dissect.evidence==3.7
dissect.executable==1.4
dissect.extfs==3.6
dissect.fat==3.7
dissect.ffs==3.6
dissect.hypervisor==3.10
dissect.ntfs==3.7
dissect.ole==3.6
dissect.regf==3.8
dissect.shellitem==3.6
dissect.sql==3.7
dissect.squashfs==1.3
dissect.target==3.13
dissect.thumbcache==1.6
dissect.util==3.12
dissect.vmfs==3.6
dissect.volume==3.7
dissect.xfs==3.6
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.0
flow.record==3.13
fusepy==3.0.1
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msgpack==1.0.7
parso==0.8.3
pexpect==4.8.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.19.0
Pygments==2.17.2
PyYAML==6.0.1
six==1.16.0
stack-data==0.6.3
structlog==23.2.0
traitlets==5.13.0
waitress==2.1.2
wcwidth==0.2.12
Werkzeug==3.0.1
yara-python==4.3.1
zstandard==0.22.0
